Split AsciidocFormatterFunc into multiple classes
